Add exception handler

+++ src/main/java/com/example/test/controller/BaseController.java
@@ -2,8 +2,12 @@ package com.example.test.controller;
 
 
 import com.example.test.entity.Book;
+import com.example.test.excepiton.NoSuchBookException;
+import com.example.test.excepiton.NotCorrectDataResponse;
 import com.example.test.service.BookService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,7 +29,10 @@ public class BaseController {
 
     @GetMapping("/books/book/{id}")
     public Book getBook(@PathVariable("id") int id){
-        return bookService.findBookById(id);
+        Book book = bookService.findBookById(id);
+        if(book==null) throw new NoSuchBookException("Book with id " + id +" not exist");
+
+        return book;
     }
 
 
@@ -44,9 +51,18 @@ public class BaseController {
 
     @DeleteMapping("/books/book/{id}")
     public String deleteBook(@PathVariable int id){
+        if(bookService.findBookById(id)==null) throw new NoSuchBookException("Book with id " + id +" not exist");
+
+
         bookService.deleteBook(id);
         return "Книга была удалена, ее id " + id;
     }
 
+    @ExceptionHandler
+    public ResponseEntity<NotCorrectDataResponse> exceptionHandle(NoSuchBookException exception){
+        NotCorrectDataResponse inf = new NotCorrectDataResponse();
+        inf.setInfo(exception.getMessage());
+        return new ResponseEntity<NotCorrectDataResponse>(inf, HttpStatus.NOT_FOUND);
+    }
 
 }
